requirements.txt
python-dotenv
pandas
numpy
matplotlib
seaborn
pymongo
certifi
pymongo[srv]
scikit-learn
dill
pyaml